fix: changed attributes in userposition and dto from int to double because of front

+++ src/Modules/Stakeholders/Explorer.Stakeholders.API/Dtos/UserPositionDto.cs
@@ -10,8 +10,8 @@ namespace Explorer.Stakeholders.API.Dtos
     {
         public int Id { get; set; }
         public long UserId { get; set; }
-        public int Latitude { get; set; }
-        public int Longitude { get; set; }
+        public double Latitude { get; set; }
+        public double Longitude { get; set; }
     }
 }
